+++ stan/math/prim/scal/fun/log.hpp
@@ -1,6 +1,7 @@
 #ifndef STAN_MATH_PRIM_SCAL_FUN_LOG_HPP
 #define STAN_MATH_PRIM_SCAL_FUN_LOG_HPP
 
+#include <stan/math/prim/meta.hpp>
 #include <cmath>
 
 namespace stan {
@@ -13,7 +14,7 @@ namespace math {
  * @param[in] x Argument.
  * @return Natural log of argument.
  */
-inline double log(int x) { return std::log(static_cast<double>(x)); }
+inline double log(int x) { return std::log(x); }
 
 }  // namespace math
 }  // namespace stan
